new market version, lots of new features

+++ navit/zipfile.h
@@ -1,4 +1,23 @@
 /**
+ * ZANavi, Zoff Android Navigation system.
+ * Copyright (C) 2011-2014 Zoff <zoff@zoff.cc>
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * version 2 as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the
+ * Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor,
+ * Boston, MA  02110-1301, USA.
+ */
+
+/**
  * Navit, a modular navigation system.
  * Copyright (C) 2005-2008 Navit Team
  *
@@ -26,24 +45,28 @@
 
 #define zip_split_sig 0x08074b50
 
-struct zip_split {
+struct zip_split
+{
 	int zipsplitsig;
 };
 
+/* this is the normal "magic" header we use!! */
 #define zip_lfh_sig 0x04034b50
+/* this is the normal "magic" header we use!! */
 
 #ifdef _MSC_VER
 #pragma pack(push,1)
 #endif /* _MSC_VER */
-struct zip_lfh {
+struct zip_lfh
+{
 	int ziplocsig;
-	short zipver; 		// 4
-	short zipgenfld;	// 6
-	short zipmthd;		// 8
-	short ziptime;		// 10
-	short zipdate;		// 12
-	int zipcrc;		// 14
-	unsigned int zipsize;  // 18
+	short zipver; // 4
+	short zipgenfld; // 6
+	short zipmthd; // 8
+	short ziptime; // 10
+	short zipdate; // 12
+	int zipcrc; // 14
+	unsigned int zipsize; // 18
 	unsigned int zipuncmp; // 22
 	unsigned short zipfnln; // 26
 	unsigned short zipxtraln; // 30
@@ -60,7 +83,8 @@ struct zip_lfh {
 #ifdef _MSC_VER
 #pragma pack(push,1)
 #endif /* _MSC_VER */
-struct zip_cd {
+struct zip_cd
+{
 	int zipcensig;
 	char zipcver;
 	char zipcos;
@@ -80,7 +104,7 @@ struct zip_cd {
 	unsigned short zipint;
 	unsigned int zipext;
 	unsigned int zipofst;
-	char zipcfn[0];	
+	char zipcfn[0];
 #ifndef _MSC_VER
 }__attribute__ ((packed));
 #else /* _MSC_VER */
@@ -88,11 +112,11 @@ struct zip_cd {
 #pragma pack(pop)
 #endif /* _MSC_VER */
 
-
 #ifdef _MSC_VER
 #pragma pack(push,1)
 #endif /* _MSC_VER */
-struct zip_cd_ext {
+struct zip_cd_ext
+{
 	short tag;
 	short size;
 	unsigned long long zipofst;
@@ -106,13 +130,14 @@ struct zip_cd_ext {
 #ifdef _MSC_VER
 #pragma pack(push,1)
 #endif /* _MSC_VER */
-struct zip_enc {
+struct zip_enc
+{
 	short efield_header;
 	short efield_size;
 	short version;
-	char vendor_id1,vendor_id2;
+	char vendor_id1, vendor_id2;
 	char encryption_strength;
-	short compress_method; 
+	short compress_method;
 #ifndef _MSC_VER
 }__attribute__ ((packed));
 #else /* _MSC_VER */
@@ -125,16 +150,17 @@ struct zip_enc {
 #ifdef _MSC_VER
 #pragma pack(push,1)
 #endif /* _MSC_VER */
-struct zip_eoc {
-	int zipesig; 		/* end of central dir signature */
+struct zip_eoc
+{
+	int zipesig; /* end of central dir signature */
 	unsigned short zipedsk; /* number of this disk */
 	unsigned short zipecen; /* number of the disk with the start of the central directory */
 	unsigned short zipenum; /* total number of entries in the central directory on this disk */
 	unsigned short zipecenn; /* total number of entries in the central directory */
-	unsigned int zipecsz; 	/* size of the central directory */
-	unsigned int zipeofst; 	/* offset of start of central directory with respect to the starting disk number */
-	short zipecoml; 	/* .ZIP file comment length */
-	char zipecom[0];	/* .ZIP file comment */
+	unsigned long long zipecsz; /* size of the central directory */
+	unsigned long long zipeofst; /* offset of start of central directory with respect to the starting disk number */
+	short zipecoml; /* .ZIP file comment length */
+	char zipecom[0]; /* .ZIP file comment */
 #ifndef _MSC_VER
 }__attribute__ ((packed));
 #else /* _MSC_VER */
@@ -147,18 +173,19 @@ struct zip_eoc {
 #ifdef _MSC_VER
 #pragma pack(push,1)
 #endif /* _MSC_VER */
-struct zip64_eoc {
-	int zip64esig;			/* zip64 end of central dir signature */
-	unsigned long long zip64esize;	/* size of zip64 end of central directory record */
-	unsigned short zip64ever;	/* version made by */
-	unsigned short zip64eneed;	/* version needed to extract */
-	unsigned int zip64edsk;		/* number of this disk */
-	unsigned int zip64ecen;		/* number of the disk with the start of the central directory */
-	unsigned long long zip64enum; 	/* total number of entries in the central directory on this disk */
-	unsigned long long zip64ecenn;	/* total number of entries in the central directory */
-	unsigned long long zip64ecsz;	/* size of the central directory */
-	unsigned long long zip64eofst;	/* offset of start of central directory with respect to the starting disk number */
-	char zip64ecom[0];		/* zip64 extensible data sector */
+struct zip64_eoc
+{
+	int zip64esig; /* zip64 end of central dir signature */
+	unsigned long long zip64esize; /* size of zip64 end of central directory record */
+	unsigned short zip64ever; /* version made by */
+	unsigned short zip64eneed; /* version needed to extract */
+	unsigned int zip64edsk; /* number of this disk */
+	unsigned int zip64ecen; /* number of the disk with the start of the central directory */
+	unsigned long long zip64enum; /* total number of entries in the central directory on this disk */
+	unsigned long long zip64ecenn; /* total number of entries in the central directory */
+	unsigned long long zip64ecsz; /* size of the central directory */
+	unsigned long long zip64eofst; /* offset of start of central directory with respect to the starting disk number */
+	char zip64ecom[0]; /* zip64 extensible data sector */
 #ifndef _MSC_VER
 }__attribute__ ((packed));
 #else /* _MSC_VER */
@@ -171,7 +198,8 @@ struct zip64_eoc {
 #ifdef _MSC_VER
 #pragma pack(push,1)
 #endif /* _MSC_VER */
-struct zip64_eocl {
+struct zip64_eocl
+{
 	int zip64lsig;
 	int zip64ldsk;
 	long long zip64lofst;
